requirements.txt
# Social Media Posting Dependencies
python-telegram-bot~=22.3
discord.py~=2.6.3
facebook-sdk==3.1.0
tweepy~=4.16.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
schedule==1.2.0
protobuf~=4.21.12
mysql-connector-python==8.2.0

# AI Content Generation
google-generativeai>=0.3.0

# Video and Audio Processing
moviepy>=1.0.3
torch>=1.9.0
torchaudio>=0.9.0

# YouTube Upload
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Additional Utilities
pathlib2>=2.3.0